+++ fv3net/regression/sklearn/__init__.py
@@ -1,1 +1,7 @@
 from .wrapper import SklearnWrapper, RegressorEnsemble
+from ._mapper import (
+    SklearnPredictionMapper,
+    PREDICT_COORD,
+    TARGET_COORD,
+    DERIVATION_DIM,
+)