+++ workflows/training_data_diags/submit_job.sh
@@ -3,7 +3,7 @@
 set -e
 
 CONFIG_FILE="workflows/training_data_diags/training_data_sources_config.yml"
-OUTPUT_PATH="gs://vcm-ml-scratch/brianh/training-data-diagnostics"
+OUTPUT_PATH="gs://vcm-ml-scratch/annak/training-data-diagnostics"
 
 
 LOGDIR="./logs"